Drop casts that only restate the destination type
Where the converted value lands directly in a declared int/size_t/etc.
(initializers, returns, plain and compound assignments to a member or
local of exactly that type), the implicit conversion says the same thing
and the cast is dropped. Each site's destination type was checked
against its declaration; casts that do real work (sign-compare
silencing, integer-vs-float division, template deduction, overload
selection, enum-to-int) stay as T(x).

+++ engine/apps/dataloader_smoke.cpp
@@ -60,7 +60,7 @@ int main(int argc, char** argv) {
     else if (a == "--prefetch")
       params.num_prefetch_threads = std::stoi(next());
     else if (a == "--budget-bytes")
-      params.memory_budget = int64_t(std::stoll(next()));
+      params.memory_budget = std::stoll(next());
     else if (a == "--budget")
       params.memory_budget = int64_t(std::stoll(next())) * 1024 * 1024;
     else if (a == "--phase") {
@@ -115,7 +115,7 @@ int main(int argc, char** argv) {
 
   // Drain one epoch using the streaming API.
   DataLoader::EpochConfig cfg;
-  cfg.batch_size = int(n_load);
+  cfg.batch_size = n_load;
   cfg.post_move = post_move;
   cfg.apply_symmetry = false;
   cfg.seed = 1;
