+++ engine/apps/move_set_eval_target_generator.cpp
@@ -88,7 +88,7 @@ struct Options {
   bool full_sweep = false;
   int sweep_cap = 1500;
   int positions_per_game = 0;  // 0 = every training-eligible turn
-  int threads = int(std::max(1u, std::thread::hardware_concurrency()));
+  int threads = std::max(1u, std::thread::hardware_concurrency());
   uint64_t seed = 0;
   int limit_games = 0;  // 0 = all games per file (a cap makes smoke runs cheap)
   // Derived in main() from the inference batch size, not a CLI flag: a slice
@@ -212,8 +212,8 @@ void encode_slices(const binlog::PositionEncoder& encoder, const GameLog& g,
     const size_t count = std::min(size_t(slice_candidates), total - first);
     CandidateSlice slice;
     slice.pos = w;
-    slice.first = uint32_t(first);
-    slice.position_candidates = uint32_t(total);
+    slice.first = first;
+    slice.position_candidates = total;
     slice.num_legal_moves = sel.num_legal_moves;
     slice.candidates.assign(sel.candidates.begin() + ptrdiff_t(first),
                             sel.candidates.begin() + ptrdiff_t(first + count));
@@ -314,7 +314,7 @@ InferenceLoop::InferenceLoop(TeacherService* service, int row_floats, int batch_
 void InferenceLoop::run(SliceQueue* queue) {
   CandidateSlice item;
   while (queue->pop(&item)) {
-    const int count = int(item.candidates.size());
+    const int count = item.candidates.size();
     if (pending_rows_ + count > batch_size_) {
       flush();
       pending_rows_ = 0;
@@ -330,13 +330,13 @@ void InferenceLoop::flush() {
   int rows = 0;
   for (const CandidateSlice& p : pending_) {
     std::memcpy(inputs_.data() + rows * row_floats_, p.rows.data(), p.rows.size() * sizeof(float));
-    rows += int(p.candidates.size());
+    rows += p.candidates.size();
   }
   float* const head_out[] = {wld_buf_.data(), score_diff_buf_.data()};
   service_->evaluate({inputs_.data(), rows}, head_out, label_planes_ ? masks_.data() : nullptr);
   int cursor = 0;
   for (CandidateSlice& p : pending_) {
-    const int count = int(p.candidates.size());
+    const int count = p.candidates.size();
     p.targets.resize(count * move_set_eval::kTargetFloatsV1);
     if (label_planes_) {
       p.planes.assign(masks_.data() + cursor * kPlaneFloats,
